Change how we enforce deny warnings
Enforce that no warning is being added as part of our regression script
instead at a crate level. This will also enforce no warnings for all
our crates, instead of just kani-compiler as is today.

This makes development a bit easier, since warnings of unused code shows
up fairly often when a feature is half implemented. So this allow us to
compile and test the code before it's production ready. That said, it
is nice to avoid warnings from piling up, so change the check to the
regression instead, which should be executed before we create PR / merge
changes.

This also aligns with the recommendation from
https://rust-unofficial.github.io/patterns/anti_patterns/deny-warnings.html