Enable `-Z stubbing` error out instead of ignore stub
Enable `-Z stubbing`, deprecate `--enable-stubbing`, and emit an error
if the user tries to verify a harness that has stubs without explicitly
enabling stubs. This used to be a warning before which was easy to miss.
It was also hard to debug when the harness would take a while to run.