Add support to `simd_bitmask` (#2677)
Create a sequential model of `simd_bitmask` and create a transformation pass that replace calls to the `simd_bitmask` intrinsic by the new model.

We will only replace those usages that we can safely get information about the `simd` representation. All other cases will be reported as unsupported feature. Type checking is not currently done for any `simd` operation and should be fixed (#2692).

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>