Unify Kani bundle CI jobs
We have two different ways today of building and testing release
bundles, one in `kani.yml` and another one in `release.yml`.

This is error prone, since the release workflow only runs during a
release, and it can diverge from what's being tested in CI.

Instead, we will always run the same workflow, except for the steps that
create the release.