Enable `-Z stubbing` and error out instead of ignore stub (#2678)
Enable `-Z stubbing`, deprecate `--enable-stubbing`, and emit an error if the user tries to verify a harness that has stubs without explicitly enabling stubs. This used to be a warning before which was easy to miss. It was also hard to debug when the harness would take a while to run.

Co-authored-by: Adrian Palacios <73246657+adpaco-aws@users.noreply.github.com>