[Boogie Backend] Use the original names of variables when generating Boogie (#2902)
For MIR variables that map to a user variable in the source code, use
their name in the source code in the generated Boogie.

If there are multiple occurrences (due to SSA), append a `_i` to the
variable name, where `i` is 0, 1, ...

This change allows one to refer to the original variable name when
adding contracts (e.g. loop invariants) in the output Boogie.

Example:

For the following Rust program:

```rust
#[kani::proof]
fn main() {
    let mut x = 41;
    let mut y = 43;
    x = 42;
    y = 42;
    kani::assert(x == y, "");
}
```
before this change, the generated Boogie was:
```
// Procedures:
procedure _RNvCsjrUp1z5o1KH_5test64main() 
{
  var _1: int;
  var _2: int;
  var _4: bool;
  var _5: int;
  var _6: int;
  _1 := 41;
  _2 := 43;
  _1 := 42;
  _2 := 42;
  _5 := _1;
  _6 := _2;
  _4 := (_5 == _6);
  assert _4;
  return;
}
```
but with this change, the generated Boogie is:
```
// Procedures:
procedure _RNvCsjrUp1z5o1KH_5test64main() 
{
  var x: int;
  var y: int;
  var _4: bool;
  var _5: int;
  var _6: int;
  x := 41;
  y := 43;
  x := 42;
  y := 42;
  _5 := x;
  _6 := y;
  _4 := (_5 == _6);
  assert _4;
  return;
}
```

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.