[Boogie Backend] Add two Boogie tests (#2914)
Add a new script-based suite for running Boogie tests. Since we don't
have Boogie integrated in our CI or the Kani driver, this suite just
checks that the generated Boogie file contains certain lines that we're
expecting.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.