[Boogie Backend] Add an unbounded array API (#2955)
This PR adds an API for creating arrays with non-deterministic content
and length. The array is implemented as a Boogie datatype with a data
field (a Boogie map), and a length field (a bitvector).

The PR also adds a test that demonstrates how the API can be used for
performing unbounded verification.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.