Fix ICE due to mismatched arguments
Use FnAbi instead of function signature when generating code for
function types. Properly check the `PassMode::Ignore`.
For foreign functions, instead of ignoring the declaration type, cast
the arguments and return value.

For now, we also ignore the caller location, since we don't currently
support tracking caller location. This change makes it easier for us to
do so. We might want to wait for this issue to get fixed so we can
easily add support using stable APIs:

https://github.com/rust-lang/project-stable-mir/issues/62