Fix ICEs due to mismatched arguments (#2994)
Use FnAbi instead of function signature when generating code for
function types. Properly check the `PassMode::Ignore`. For foreign
functions, instead of ignoring the declaration type, cast the arguments
and return value.

For now, we also ignore the caller location, since we don't currently
support tracking caller location. This change makes it easier for us to
do so. We might want to wait for this issue to get fixed so we can
easily add support using stable APIs:
https://github.com/rust-lang/project-stable-mir/issues/62

Resolves #2260
Resolves #2312
Resolves #1365
Resolves #1350