Add `--use-local-toolchain` to Kani setup (#3056)
Adds `--use-local-toolchain` to Kani's setup flow, which accepts a local
toolchain and then uses that to finish the Kani setup.

Some notes:

1. Why? This is mainly for installing GPG verified toolchains. 
2. This is missing some cleanup and refactoring work, like ensuring that
the user defined toolchain matches the one that Kani was built against
etc. Marked as Todo, for later.

Resolves [#3058](https://github.com/model-checking/kani/issues/3058)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.