Upgrade toolchain to 2024-03-11 (#3071)
Relevant upstream changes:

https://github.com/rust-lang/rust/pull/120675: An intrinsic `Symbol` is
now wrapped in a `IntrinsicDef` struct, so the relevant part of the code
needed to be updated.
https://github.com/rust-lang/rust/pull/121464: The second argument of
the `create_wrapper_file` function changed from a vector to a string.
https://github.com/rust-lang/rust/pull/121662: `NullOp::DebugAssertions`
was renamed to `NullOp::UbCheck` and it now has data (currently unused
by Kani)
https://github.com/rust-lang/rust/pull/121728: Introduces `F16` and
`F128`, so needed to add stubs for them
https://github.com/rust-lang/rust/pull/121969: `parse_sess` was renamed
to `psess`, so updated the relevant code.
https://github.com/rust-lang/rust/pull/122059: The
`is_val_statically_known` intrinsic is now used in some `core::fmt`
code, so had to handle it in (codegen'ed to false).
https://github.com/rust-lang/rust/pull/122233: This added a new
`retag_box_to_raw` intrinsic. This is an operation that is primarily
relevant for stacked borrows. For Kani, we just return the pointer.

Resolves #3057