Add CI test for --use-local-toolchain (#3074)
Adds a CI test for --use-local-toolchain.
This test will run after every push to main.

Additional cleanup to the setup as mentioned in #3060 in next PR's. (In
interest of keeping changes small & focused).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.