Allow modifies clause for verification only (#3098)
Allow contracts to be used for verification, even if it is not suitable
for stubbing. For that, we remove the requirement that modifies and
return types of a function annotated with contracts must implement
`kani::Arbitrary`, since that is only needed for recursion and stubbing
with contract.

This is done via a new intrinsic `any_modifies` to Kani that should only
be used by contract instrumentation. The `T: Arbitrary` requirement is
only checked when users try to use the contract as stub or to check
recursive functions.

For now, we also require users to annotate their contracts with
`kani::recursion` if they want to use inductive reasoning to verify a
recursive function.

Resolves https://github.com/model-checking/kani/issues/2997.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Signed-off-by: Felipe R. Monteiro <felisous@amazon.com>