Remove unnecessary build step for some workflows (#3124)
Workflows that run the perf regressions (`scripts/kani-perf.sh`) do not
need to build Kani, as the perf script itself builds Kani in release
mode:


https://github.com/model-checking/kani/blob/dc0978043c52492112e4ad37a617fd3c8100ef1f/scripts/kani-perf.sh#L12

The extra build step was causing Kani to be built two extra times for
the `benchcomp` flow because the workflow was building it in debug mode,
so the script will end up rebuilding it in release mode. This should
save about 3 minutes for the `benchcomp` flow.