Remove empty box creation from contracts impl
There seems to be an issue in CBMC contracts implementation that it
assumes that `free` must have a body. However, slicing can remove
`free` body if the harness does not allocate anything.

We used to create an empty Box before to force `free` to be in scope.
Instead, just invoke `free(NULL)` which is a no-op.